requirements.txt
gcore
pyyaml
requests
